[Transaction] Revamping Transaction Function #1

+++ library/transaction/history.c
@@ -1,6 +1,6 @@
 #include "../transaction/trxDb.h"
 
-int trxHistory(int uid)
+int trxHistory(long uid)
 {
     int choice;
     history=fopen(dbHistory, "a");
